Refactor build scripts
Our monolithic build script assumed that we wanted to build all of
the images at once. And while this was true at the beginning where we
only had a few images to maintain, this is no longer true now.

To make maintaining things a bit easier, we removed most of the
functionality from the build.sh script and only left enough pieces in
for it to be able to publish one image. Iterating over all image
definitions in now done in the build-all.sh script.